Replace VOLUME_DOWN with VOLUME_UP in the keylayouts
This effectively changes the volume down button into a second volume up
button. This sounds insane, but my user needs this... I'll revisit this
decision when/if she moves to another model of phone, but it appears this
will be necessary as long as she's using a Motorola Defy.

(I could change the keylayouts directly on the phone itself, but it's
easier for me to roll it into my custom build.)